+++ R/make-classification-fast.R
@@ -21,20 +21,27 @@
 #' split supported by `rsample`
 #' @param .split_args The default is NULL, when NULL then the default parameters
 #' of the split type will be executed for the rsample split type.
+#' @param .drop_na The default is TRUE, which will drop all NA's from the data.
 #'
 #' @examples
-#' library(recipes, quietly = TRUE)
-#' library(dplyr, quietly = TRUE)
+#' library(recipes)
+#' library(dplyr)
+#' library(tidyr)
 #'
-#' df <- mtcars |> mutate(cyl = as.factor(cyl))
-#' rec_obj <- recipe(cyl ~ ., data = df)
+#' df <- Titanic |>
+#'  as_tibble() |>
+#'  uncount(n) |>
+#'  mutate(across(everything(), as.factor))
+#'
+#' rec_obj <- recipe(Survived ~ ., data = df)
 #'
 #' fct_tbl <- fast_classification(
 #'   .data = df,
 #'   .rec_obj = rec_obj,
-#'   .parsnip_eng = c("glm","LiblineaR"))
+#'   .parsnip_eng = c("glm","earth")
+#'   )
 #'
-#' glimpse(fct_tbl)
+#' fct_tbl
 #'
 #' @return
 #' A list or a tibble.
@@ -48,7 +55,7 @@ NULL
 
 fast_classification <- function(.data, .rec_obj, .parsnip_fns = "all",
                                 .parsnip_eng = "all", .split_type = "initial_split",
-                                .split_args = NULL){
+                                .split_args = NULL, .drop_na = TRUE){
 
   # Tidy Eval ----
   call <- list(.parsnip_fns) |>
@@ -92,13 +99,17 @@ fast_classification <- function(.data, .rec_obj, .parsnip_fns = "all",
       pred_wflw = internal_make_wflw_predictions(mod_fitted_tbl, splits_obj)
     )
 
+  if (.drop_na){
+    mod_pred_tbl <- mod_pred_tbl[!sapply(mod_pred_tbl$fitted_wflw, function(x) length(x) == 0), ]
+  }
 
   # Return ----
-  class(mod_tbl) <- c("fst_reg_tbl", class(mod_tbl))
-  attr(mod_tbl, ".parsnip_engines") <- .parsnip_eng
-  attr(mod_tbl, ".parsnip_functions") <- .parsnip_fns
-  attr(mod_tbl, ".split_type") <- .split_type
-  attr(mod_tbl, ".split_args") <- .split_args
+  class(mod_pred_tbl) <- c("fst_class_tbl", class(mod_pred_tbl))
+  attr(mod_pred_tbl, ".parsnip_engines") <- .parsnip_eng
+  attr(mod_pred_tbl, ".parsnip_functions") <- .parsnip_fns
+  attr(mod_pred_tbl, ".split_type") <- .split_type
+  attr(mod_pred_tbl, ".split_args") <- .split_args
+  attr(mod_pred_tbl, ".drop_na") <- .drop_na
 
   return(mod_pred_tbl)
 }